Move function calls into the correct service files
