+++ server/service/software_installers.go
@@ -583,6 +583,10 @@ func (svc *Service) InstallVPPAppPostValidation(ctx context.Context, host *fleet
 	return "", fleet.ErrMissingLicense // called downstream of auth checks so doesn't need skipauth
 }
 
+////////////////////////////////////////////////////////////////////////////////
+// Uninstall software
+////////////////////////////////////////////////////////////////////////////////
+
 type uninstallSoftwareRequest struct {
 	HostID          uint `url:"host_id"`
 	SoftwareTitleID uint `url:"software_title_id"`
@@ -604,6 +608,10 @@ func (svc *Service) UninstallSoftwareTitle(ctx context.Context, _ uint, _ uint) 
 	svc.authz.SkipAuthorization(ctx)
 	return fleet.ErrMissingLicense
 }
+
+////////////////////////////////////////////////////////////////////////////////
+// Get software uninstall results (host details and self service)
+////////////////////////////////////////////////////////////////////////////////
 
 type getSoftwareInstallResultsRequest struct {
 	InstallUUID string `url:"install_uuid"`
@@ -653,6 +661,43 @@ func getSoftwareInstallResultsEndpoint(ctx context.Context, request interface{},
 }
 
 func (svc *Service) GetSoftwareInstallResults(ctx context.Context, resultUUID string) (*fleet.HostSoftwareInstallerResult, error) {
+	// skipauth: No authorization check needed due to implementation returning
+	// only license error.
+	svc.authz.SkipAuthorization(ctx)
+
+	return nil, fleet.ErrMissingLicense
+}
+
+////////////////////////////////////////////////////////////////////////////////
+// Get software uninstall results from My device page
+////////////////////////////////////////////////////////////////////////////////
+
+type getDeviceSoftwareUninstallResultsRequest struct {
+	Token       string `url:"token"`
+	ExecutionID string `url:"execution_id"`
+}
+
+func (r *getDeviceSoftwareUninstallResultsRequest) deviceAuthToken() string {
+	return r.Token
+}
+
+func getDeviceSoftwareUninstallResultsEndpoint(ctx context.Context, request interface{}, svc fleet.Service) (fleet.Errorer, error) {
+	host, ok := hostctx.FromContext(ctx)
+	if !ok {
+		err := ctxerr.Wrap(ctx, fleet.NewAuthRequiredError("internal error: missing host from request context"))
+		return getSoftwareInstallResultsResponse{Err: err}, nil
+	}
+
+	req := request.(*getDeviceSoftwareUninstallResultsRequest)
+	scriptResult, err := svc.GetSelfServiceUninstallScriptResult(ctx, host, req.ExecutionID)
+	if err != nil {
+		return getScriptResultResponse{Err: err}, nil
+	}
+
+	return setUpGetScriptResultResponse(scriptResult), nil
+}
+
+func (svc *Service) GetSelfServiceUninstallScriptResult(ctx context.Context, host *fleet.Host, execID string) (*fleet.HostScriptResult, error) {
 	// skipauth: No authorization check needed due to implementation returning
 	// only license error.
 	svc.authz.SkipAuthorization(ctx)
